rm new test file integrated into old
